requirements.txt
PyPDF2==3.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
numpy>=1.26.0
pandas>=2.1.0
tqdm==4.65.0
requests==2.31.0
python-json-logger==2.0.7
diskcache==5.6.3
scikit-learn>=1.3.2
torch>=2.1.0
transformers>=4.35.0
nltk==3.8.1
spacy>=3.7.0
textstat==0.7.3
arxiv==1.4.8
joblib==1.3.2
